move frame_additional_data & related into own header
